Nested hybrid grid: extract result helpers into RigNestedHybridGridResultTools
