+++ ApplicationLibCode/ReservoirDataModel/CMakeLists_files.cmake
@@ -88,6 +88,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigUniformRefinement.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigNonUniformRefinement.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigNestedHybridGridReconstructor.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RigNestedHybridGridResultTools.cpp
 )
 
 list(APPEND CODE_SOURCE_FILES ${SOURCE_GROUP_SOURCE_FILES})
